Actually commit code changes, and also performance improvements
R3Test.cs and Search.cs somehow missed the boat in the last commit, so
are here again.  They are also updated for performance improvements.
Added Pair.cs which is like KeyValuePair only it supports IEquatable<>,
so it is usable as a dictionary key (or LookupHeap value) using default
equality comparer without invoking terrible performance.  It comes with
an Item1Comparer which is like TMD.Algo.Algorithms.Generic.KeyComparer
for KeyValuePair.  Search.Pfs now uses this internally, resulting in a
15x performance improvement for 2008R3Q2Large.  It is also used in a
number of Graph algorithms, for which I did not test the performance
improvement.